requirements.txt
-i https://pypi.org/simple
attrs==23.2.0 ; python_version >= '3.7'
certifi==2024.6.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
h11==0.14.0 ; python_version >= '3.7'
idna==3.7 ; python_version >= '3.5'
outcome==1.3.0.post0 ; python_version >= '3.7'
packaging==24.1 ; python_version >= '3.8'
pysocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3 ; python_version >= '3.8'
selenium==4.22.0
sniffio==1.3.1 ; python_version >= '3.7'
sortedcontainers==2.4.0
trio==0.25.1 ; python_version >= '3.8'
trio-websocket==0.11.1 ; python_version >= '3.7'
typing-extensions==4.12.2 ; python_version >= '3.8'
urllib3==1.26.16
webdriver-manager==4.0.1
websocket-client==1.8.0 ; python_version >= '3.8'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
